requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.2
pydantic-settings==2.1.0
alembic==1.12.1
argon2-cffi
mysql-connector-python
